moving Badge props ts to types.ts
